+++ .gitignore
@@ -27,5 +27,3 @@ src/components/base-component-jade/_package
 
 test/fixtures/test.tar.gz
 test/fixtures/targz-test
-
-coverage
